Raspador de articulos inverso (2)/requirements.txt
google-cloud-vision
asana
validators